+++ R/solve_MDP.R
@@ -2,34 +2,65 @@
 
 #' Solve an MDP Problem
 #'
-#' A simple implementation of value iteration and modified policy iteration.
+#' Implementation of value iteration, modified policy iteration and other
+#' methods based on reinforcement learning techniques to solve finite
+#' state space MDPs.
+#'
+#' Implemented are the following dynamic programming methods:
+#'
+#' * **Modified Policy Iteration**
+#' starts with a random policy and iteratively performs
+#' a sequence of
+#'   - approximate policy evaluation (estimate the value function for the
+#' current policy using `k_backups`), and
+#'   - policy improvement (calculate a greedy policy given the value function).
+#' The algorithm stops when it converges to a stable policy.
+#'
+#' * **Value Iteration** starts with
+#'  an arbitraty value function (we use all 0s) and iteratively
+#' updates the value function using the Bellman equation. The iterations
+#' are terminated when no state value in the value function changes by more than
+#' \eqn{\epsilon (1-\gamma) / \gamma} which means that no state value is more than
+#' \eqn{\epsilon} from the value in the optimal value function. The greedy policy
+#' is calculated from the value function. Value iteration can be seen as
+#' policy iteration with truncated policy evaluation.
+#'
+#' Implemented are the following temporal difference control methods. Note
+#' that the MDP transition and reward models are only used to simulate
+#' the environment for these reinforcement learning methods.
+#' The algorithms use a step size parameter \eqn{\alpha} (learning rate) for the
+#' updates and the exploration parameter \eqn{\epsilon} for
+#' the \eqn{\epsilon}-greedy policy.
+#'
+#' * **Q-Learning** is an off-policy temporal difference method that uses
+#'    an \eqn{\epsilon}-greedy behavior policy and learns a greedy target
+#'    policy.
+#'
+#' * **Sarsa** is an on-policy method that learns an \eqn{\epsilon}-greedy
+#'    policy.
+#'
+#' * **Expected Sarsa**: We implement an on-policy version that uses
+#'   the expected value under the current policy for the update.
+#'   It moves deterministically in the same direction as Sarsa
+#'   moved in expectation. Because it uses the expectation, we can
+#'   set the step size \eqn{\alpha} to large values and even 1.
+#'
 #'
 #' @family solver
 #' @family MDP
 #'
-#' @param model a POMDP problem specification created with [POMDP()].
-#'   Alternatively, a POMDP file or the URL for a POMDP file can be specified.
-#' @param method string; one of the following solution methods: `'value'`,
-#'   `'policy'`.
+#' @param model an MDP problem specification.
+#' @param method string; one of the following solution methods: `'value_iteration'`,
+#'   `'policy_iteration'`, `'q_learning'`, `'sarsa'`, or `'expected_sarsa'`.
 #' @param horizon an integer with the number of epochs for problems with a
 #'   finite planning horizon. If set to `Inf`, the algorithm continues
 #'   running iterations till it converges to the infinite horizon solution. If
-#'   `NULL`, then the horizon specified in `model` will be used.  For
-#'   time-dependent POMDPs a vector of horizons can be specified (see Details
-#'   section).
-#' @param discount discount factor in range \eqn{[0, 1]}. If `NULL`, then the
+#'   `NULL`, then the horizon specified in `model` will be used.
+#' @param discount discount factor in range \eqn{(0, 1]}. If `NULL`, then the
 #'   discount factor specified in `model` will be used.
-#' @param terminal_values a vector with terminal utilities for each state. If
-#'   `NULL`, then a vector of all 0s is used.
-#' @param eps maximum error allowed in the utility of any state
-#'   (i.e., the maximum policy loss).
-#' @param max_iterations maximum number of iterations allowed to converge. If the
-#'   maximum is reached then the non-converged solution is returned with a
-#'   warning.
-#' @param k_backups number of look ahead steps used for approximate policy evaluation
-#'   used by method `'policy'`.
 #' @param verbose logical, if set to `TRUE`, the function provides the
-#'   output of the pomdp solver in the R console.
+#'   output of the solver in the R console.
+#' @param ... further parameters are passed on to the solver function.
 #'
 #' @return `solve_MDP()` returns an object of class POMDP which is a list with the
 #'   model specifications (`model`), the solution (`solution`).
@@ -39,35 +70,39 @@
 #'   - `delta` final delta (infinite-horizon only)
 #'   - `iterations` number of iterations to convergence (infinite-horizon only)
 #'
-#'
 #' @author Michael Hahsler
 #' @examples
 #' data(Maze)
 #' Maze
 #'
 #' # use value iteration
-#' maze_solved <- solve_MDP(Maze, method = "value")
+#' maze_solved <- solve_MDP(Maze, method = "value_iteration")
+#' maze_solved
 #' policy(maze_solved)
 #'
-#' # value function (utility function U)
+#' # plot the value function U (with the state labels rotated)
 #' plot_value_function(maze_solved)
 #'
-#' # Q-function (states times action)
-#' q_values_MDP(maze_solved)
+#' # Maze solutions can be visualized
+#' plot_Maze_solution(maze_solved)
 #'
 #' # use modified policy iteration
-#' maze_solved <- solve_MDP(Maze, method = "policy")
+#' maze_solved <- solve_MDP(Maze, method = "policy_iteration")
 #' policy(maze_solved)
 #'
 #' # finite horizon
-#' maze_solved <- solve_MDP(Maze, method = "value", horizon = 3)
+#' maze_solved <- solve_MDP(Maze, method = "value_iteration", horizon = 3)
 #' policy(maze_solved)
+#' plot_Maze_solution(maze_solved, epoch = 1)
+#' plot_Maze_solution(maze_solved, epoch = 2)
+#' plot_Maze_solution(maze_solved, epoch = 3)
 #'
 #' # create a random policy where action n is very likely and approximate
 #' #  the value function. We change the discount factor to .9 for this.
 #' Maze_discounted <- Maze
 #' Maze_discounted$discount <- .9
-#' pi <- random_MDP_policy(Maze_discounted, prob = c(n = .7, e = .1, s = .1, w = 0.1))
+#' pi <- random_MDP_policy(Maze_discounted, 
+#'         prob = c(n = .7, e = .1, s = .1, w = 0.1))
 #' pi
 #'
 #' # compare the utility function for the random policy with the function for the optimal
@@ -79,20 +114,56 @@
 #'
 #' # Note that the solver already calculates the utility function and returns it with the policy
 #' policy(maze_solved)
+#'
+#' # Learn a Policy using Q-Learning
+#' maze_learned <- solve_MDP(Maze, method = "q_learning", N = 100)
+#' maze_learned
+#'
+#' maze_learned$solution
+#' policy(maze_learned)
+#' plot_value_function(maze_learned)
+#' plot_Maze_solution(maze_learned)
 #' @export
-solve_MDP <- function(model,
-  horizon = NULL,
-  discount = NULL,
-  terminal_values = NULL,
-  method = "value",
-  eps = 0.01,
-  max_iterations = 1000,
-  k_backups = 10,
-  verbose = FALSE) {
+solve_MDP <- function(model, method = "value", ...) {
+  methods_DP <- c("value_iteration", "policy_iteration")
+  methods_TD <-  c("sarsa", "q_learning", "expected_sarsa")
+  
+  method <- match.arg(method, c(methods_DP, methods_TD))
+  
+  if (method %in% methods_DP)
+    return(solve_MDP_DP(model, method, ...))
+  if (method %in% methods_TD)
+    return(solve_MDP_TD(model, method, ...))
+  
+  # we should not get here!
+  stop("Unknown method!")
+}
+
+#' @rdname solve_MDP
+#' @param terminal_values a vector with terminal utilities for each state. If
+#'   `NULL`, then a vector of all 0s is used.
+#' @param eps maximum error allowed in the utility of any state
+#'   (i.e., the maximum policy loss) used as the termination criterion for
+#'   the value iteration method.
+#' @param max_iterations maximum number of iterations allowed to converge. If the
+#'   maximum is reached then the non-converged solution is returned with a
+#'   warning.
+#' @param k_backups number of look ahead steps used for approximate policy evaluation
+#'   used by the policy iteration method.
+#' @export
+solve_MDP_DP <- function(model,
+                         method = "value_iteration",
+                         horizon = NULL,
+                         discount = NULL,
+                         terminal_values = NULL,
+                         eps = 0.01,
+                         max_iterations = 1000,
+                         k_backups = 10,
+                         verbose = FALSE) {
   if (!inherits(model, "MDP"))
     stop("'model' needs to be of class 'MDP'.")
   
-  methods <- c("value", "policy")
+  methods <- c("value_iteration", "policy_iteration")
   method <- match.arg(method, methods)
   
   ### default is infinite horizon
@@ -105,35 +176,33 @@ solve_MDP <- function(model,
     model$discount <- discount
   if (is.null(model$discount)) {
     message("No discount rate specified. Using .9!")
-    model$discount <- .9
+    model$discount <- .10
   }
   
-  if (is.infinite(model$horizon))
-    switch(
-      method,
-      value = MDP_value_iteration_inf_horizon(model, eps, max_iterations,
-        U = terminal_values, verbose = verbose),
-      policy = MDP_policy_iteration_inf_horizon(
-        model,
-        eps,
-        max_iterations,
-        k_backups,
-        U = terminal_values,
-        verbose = verbose
-      )
-    )
-  else
-    switch(
-      method,
-      value = MDP_value_iteration_finite_horizon(
-        model,
-        horizon = model$horizon,
-        U = terminal_values,
-        verbose = verbose
-      ),
-      policy = stop("Method not implemented yet for finite horizon problems.")
-    )
-  
+  switch(method,
+         value_iteration = {
+           if (is.infinite(model$horizon))
+             MDP_value_iteration_inf_horizon(model,
+                                             eps,
+                                             max_iterations,
+                                             U = terminal_values,
+                                             verbose = verbose)
+           else
+             MDP_value_iteration_finite_horizon(model,
+                                                horizon = model$horizon,
+                                                U = terminal_values,
+                                                verbose = verbose)
+         },
+         policy_iteration = {
+           if (is.infinite(model$horizon))
+             MDP_policy_iteration_inf_horizon(model,
+                                              max_iterations,
+                                              k_backups,
+                                              U = terminal_values,
+                                              verbose = verbose)
+           else
+             stop("Method not implemented yet for finite horizon problems.")
+         })
 }
 
 # Q-Function:
@@ -142,48 +211,22 @@ solve_MDP <- function(model,
 .QV <-
   function(s, a, P, R, GAMMA, U)
     sum(P[[a]][s,] * (R[[a]][[s]] + GAMMA * U))
-
 .QV_vec <- Vectorize(.QV, vectorize.args = c("s", "a"))
-
-#' @rdname solve_MDP
-#' @param U a vector with state utilities (expected sum of discounted rewards from that point on).
-#' @return `q_values_MDP()` returns a state by action matrix specifying the Q-function,
-#'   i.e., the utility value of executing each action in each state.
-#' @export
-q_values_MDP <- function(model, U = NULL) {
-  if (!inherits(model, "MDP"))
-    stop("'model' needs to be of class 'MDP'.")
-  
-  S <- model$states
-  A <- model$actions
-  P <- transition_matrix(model, sparse = TRUE)
-  # FIXME: sparse = TRUE returns a dataframe
-  R <- reward_matrix(model, sparse = FALSE)
-  policy <- model$solution$policy[[1]]
-  GAMMA <- model$discount
-  
-  if (is.null(U))
-    if (!is.null(policy))
-      U <- policy$U
-  else
-    stop("'model' does not contain state utilities (it is unsolved). You need to specify U.")
-  
-  structure(outer(S, A, .QV_vec, P, R, GAMMA, U), dimnames = list(S, A))
-}
 
 # TODO: we could check for convergence
 MDP_value_iteration_finite_horizon <-
   function(model,
-    horizon,
-    U = NULL,
-    verbose = FALSE) {
+           horizon,
+           U = NULL,
+           verbose = FALSE) {
     if (!inherits(model, "MDP"))
       stop("'model' needs to be of class 'MDP'.")
     
     S <- model$states
     A <- model$actions
     P <- transition_matrix(model, sparse = TRUE)
-    R <- reward_matrix(model, sparse = FALSE) ## Note sparse leaves it as a data.frame
+    R <-
+      reward_matrix(model, sparse = FALSE) ## Note sparse leaves it as a data.frame
     GAMMA <- model$discount
     
     horizon <- as.integer(horizon)
@@ -213,16 +256,16 @@ MDP_value_iteration_finite_horizon <-
     }
     
     model$solution <- list(policy = policy,
-      converged = NA)
+                           converged = NA)
     model
   }
 
 MDP_value_iteration_inf_horizon <-
   function(model,
-    eps,
-    max_iterations = 1000,
-    U = NULL,
-    verbose = FALSE) {
+           eps,
+           max_iterations = 1000,
+           U = NULL,
+           verbose = FALSE) {
     S <- model$states
     A <- model$actions
     P <- transition_matrix(model)
@@ -284,75 +327,12 @@ MDP_value_iteration_inf_horizon <-
     model
   }
 
-# Policy iteration with approximate policy evaluation
-
-#' @rdname solve_MDP
-#' @param prob probability vector for actions.
-#' @return `random_MDP_policy()` returns a data.frame with columns state and action to define a policy.
-#' @export
-random_MDP_policy <-
-  function(model, prob = NULL) {
-    if (!inherits(model, "MDP"))
-      stop("'model' needs to be of class 'MDP'.")
-    
-    A <- model$actions
-    S <- model$states
-    
-    data.frame(
-      state = model$states,
-      action = factor(
-        sample(
-          seq_along(A),
-          size = length(S),
-          replace = TRUE,
-          prob = prob
-        ),
-        levels = seq_along(A),
-        labels = A
-      )
-    )
-  }
-
-#' @rdname solve_MDP
-#' @param pi a policy as a data.frame with columns state and action.
-#' @return `approx_MDP_policy_evaluation()` is used by the modified policy
-#'   iteration algorithm and returns an approximate utility vector U estimated by evaluating policy `pi`.
-#' @export
-approx_MDP_policy_evaluation <-
-  function(pi,
-    model,
-    U = NULL,
-    k_backups = 10) {
-    if (!inherits(model, "MDP"))
-      stop("'model' needs to be of class 'MDP'.")
-    
-    S <- model$states
-    A <- model$actions
-    P <- transition_matrix(model, sparse = TRUE)
-    R <- reward_matrix(model, sparse = FALSE)
-    GAMMA <- model$discount
-    
-    if (is.data.frame(pi))
-      pi <- pi$action
-    names(pi) <- S
-    
-    # start with all 0s if no previous U is given
-    if (is.null(U))
-      U <- rep(0, times = length(S))
-    
-    for (i in seq_len(k_backups))
-      U <- .QV_vec(S, pi, P, R, GAMMA, U)
-    
-    U
-  }
-
 MDP_policy_iteration_inf_horizon <-
   function(model,
-    eps,
-    max_iterations = 1000,
-    k_backups = 10,
-    U = NULL,
-    verbose = FALSE) {
+           max_iterations = 1000,
+           k_backups = 10,
+           U = NULL,
+           verbose = FALSE) {
     S <- model$states
     A <- model$actions
     P <- transition_matrix(model, sparse = TRUE)
@@ -406,5 +386,127 @@ MDP_policy_iteration_inf_horizon <-
       converged = converged,
       iterations = i
     )
+    model
+  }
+
+#' @rdname solve_MDP
+#' @param alpha step size in `(0, 1]`.
+#' @param epsilon used for \eqn{\epsilon}-greedy policies.
+#' @param N number of episodes used for learning.
+#' @export
+solve_MDP_TD <-
+  function(model,
+           method = "q_learning",
+           horizon = NULL,
+           discount = NULL,
+           alpha = 0.1,
+           epsilon = 0.1,
+           N = 100,
+           verbose = FALSE) {
+    ### default is infinite horizon, but we use 1000 to guarantee termination
+    warn_horizon <- FALSE
+    if (is.null(horizon))
+      horizon <- model$horizon
+    if (is.null(horizon)) {
+      warn_horizon <- TRUE
+      horizon <- 10000
+    }
+    
+    if (is.null(discount))
+      discount <- model$discount
+    if (is.null(discount))
+      discount <- 1
+    gamma <- discount
+    model$discount <- discount
+    
+    S <- model$states
+    S_absorbing <- S[which(absorbing_states(model))]
+    A <- model$actions
+    P <- transition_matrix(model, sparse = TRUE)
+    #R <- reward_matrix(model, sparse = FALSE)
+    start <- .translate_belief(NULL, model = model)
+    
+    method <-
+      match.arg(method, c("sarsa", "q_learning", "expected_sarsa"))
+    
+    # Initialize Q
+    Q <-
+      matrix(0,
+             nrow = length(S),
+             ncol = length(A),
+             dimnames = list(S, A))
+    
+    # loop episodes
+    for (e in seq(N)) {
+      s <- sample(S, 1, prob = start)
+      a <- greedy_MDP_action(s, Q, epsilon)
+      
+      # loop steps in episode
+      i <- 1L
+      while (TRUE) {
+        s_prime <- sample(S, 1L, prob = P[[a]][s,])
+        r <- reward_val(model, a, s, s_prime)
+        a_prime <- greedy_MDP_action(s_prime, Q, epsilon)
+        
+        if (verbose) {
+          if (i == 1L)
+            cat("\n*** Episode", e, "***\n")
+          cat("Step", i, "- s a r s' a':", s, a, r, s_prime, a_prime, "\n")
+        }
+        
+        if (method == "sarsa")
+          # is called Sarsa because it uses the sequence s, a, r, s', a'
+          Q[s, a] <-
+          Q[s, a] + alpha * (r + gamma * Q[s_prime, a_prime] - Q[s, a])
+        else if (method == "q_learning") {
+          # a' is greedy instead of using the behavior policy
+          a_max <- greedy_MDP_action(s_prime, Q, epsilon = 0)
+          Q[s, a] <-
+            Q[s, a] + alpha * (r + gamma * Q[s_prime, a_max] - Q[s, a])
+        } else if (method == "expected_sarsa") {
+          p <-
+            greedy_MDP_action(s_prime, Q, epsilon, prob = TRUE)
+          exp_Q_prime <-
+            sum(greedy_MDP_action(s_prime, Q, epsilon, prob = TRUE) * Q[s_prime,],
+                na.rm = TRUE)
+          Q[s, a] <-
+            Q[s, a] + alpha * (r + gamma * exp_Q_prime - Q[s, a])
+        }
+        
+        s <- s_prime
+        a <- a_prime
+        
+        if (s %in% S_absorbing)
+          break
+        
+        if (i >= horizon) {
+          if (warn_horizon)
+            warning(
+              "Max episode length of ",
+              i,
+              " reached without reaching an absorbing state! Terminating episode."
+            )
+          break
+        }
+        
+        i <- i + 1L
+      }
+    }
+    
+    model$solution <- list(
+      method = method,
+      alpha = alpha,
+      epsilon = epsilon,
+      N = N,
+      Q = Q,
+      converged = NA,
+      policy = list(data.frame(
+        state = S,
+        U = apply(Q, MARGIN = 1, max),
+        action = A[apply(Q, MARGIN = 1, which.max)],
+        row.names = NULL
+      ))
+    )
+    
     model
   }