+++ NEWS.md
@@ -1,7 +1,13 @@
-# pomdp 1.1.3.1 (xx/xx/2023)
+# pomdp 1.1.3.1 (xx/xx/2024)
+
+## New Features
+* Policies for the Maze problem can now be visualized
+* Added functions to work with MDP policies (see ? MDP_policy_functions).
+* Added MDP solver functions: Q-leanring, Sarsa, and expected Sarsa
 
 ## Changes
 * simulate_POMDP(): Better calculation of T for infinite-horizon problems.
+* several functions are now generics with methods for POMDP and MDP.
 
 ## Bugfixes
 * q_values_MDP(), solve_MDP(): Fixed reward representation issue.